fix acc

+++ konchev.maksim/T2/DataStruct.h
@@ -6,7 +6,7 @@ namespace konchev
 {
   struct DataStruct
   {
-    long long key1;
+    int key1;
     unsigned long long key2;
     std::string key3;
   };
